requirements.txt
streamlit>=1.22.0
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python>=4.8.0.74
numpy>=1.25.0
